Stage 1: batched fp32 answer log-prob scoring, score distribution plot, threshold filter script

+++ setup.py
@@ -2,7 +2,7 @@ import torch
 from transformers import AutoModelForCausalLM, AutoTokenizer
 
 
-def load_model(model_id="Qwen/Qwen3.5-4B", dtype=torch.bfloat16, device="cuda"):
+def load_model(model_id="Qwen/Qwen3.5-4B", dtype=torch.float32, device="cuda"):
     tok = AutoTokenizer.from_pretrained(model_id)
     model = AutoModelForCausalLM.from_pretrained(model_id, dtype=dtype, device_map=device).eval()
     return model, tok
